fix(portability): address third-party Windows/macOS audit findings
P0:
- tun_mac.c: create utun via the kernel control socket
  (PF_SYSTEM/CTLIOCGINFO/connect + UTUN_OPT_IFNAME) — /dev/utunN
  devfs nodes no longer exist on modern XNU, so macOS TUN mode was
  completely broken (untested by CI, which only ran --help)
- macos CI: both arches now build a static no-shared OpenSSL from
  source — the arm64 release previously linked Homebrew dylibs from
  /opt/homebrew/opt/openssl@3 and could not run on stock macOS
- release job: chmod 0755 the macOS binaries before zipping (the
  v0.6.0 zips shipped 0644, unpacking to Permission denied) and add
  an unzip -Z exec-bit gate; Windows zips now carry README/LICENSE/
  ci/WINTUN.txt (wintun.dll arch + download note)

P1:
- relay_proxy.c: %llu + explicit casts instead of %zu; proxy.c: bound
  the snprintf %s with %.45s — newer MinGW-w64 (mingw-w64 runtime 13,
  Ubuntu 26.04 package) rejects both under -Werror. CI was green only
  because the runner image is pinned to older packages; both toolchain
  generations now pass
- port.c: win_join_argv converts UTF-8 -> UTF-16 with
  MultiByteToWideChar (byte-cast garbled netsh interface names like
  以太网, failing the server-pin route and rolling back TUN startup)
- route.c: Windows ipv6 ULA now gets the explicit /96 (netsh default
  prefix would break client-pool on-link semantics); route_teardown6
  also deletes the ULA address (wintun adapters persist across runs)
- tun_win.c: implement tun_ifname / tun_pool_write_fd /
  tun_pool_note_stall to complete the tun.h contract (latent link
  failures on any future reuse)
- CI coverage: SOCKS5 handshake suite now runs on macOS and under
  wine against the Windows build; PE import-table gate (no
  libssl/libcrypto/winpthread DLLs), ad-hoc codesign on both macOS
  arches, runner images pinned to ubuntu-24.04, llvm-mingw pinned to
  20260616
- https.c: prefer Homebrew CA bundles on macOS (the /etc/ssl/cert.pem
  symlink is stale); oidc_config.c: accept backslash path separators
  for Windows config dirs
- README: Windows cross-build instructions, macOS 11+ requirement,
  Gatekeeper/notarization notes, wintun.dll arch table; honest note
  that Windows netsh routing awaits real-machine validation

Verified locally: mingw x86_64 + i686 -Werror full build, wine SOCKS
suite (no-token + token), native SOCKS suite, Linux build, sudo
integration suite all pass.

+++ src/common/relay_proxy.c
@@ -674,11 +674,14 @@ static void *rp_dir_main(void *ud)
             prof_print(up_dir ? "rp up send" : "rp dn send", &pst2,
                        up_dir ? g_prof_rp_up_send : g_prof_rp_dn_send);
             prof_print("rp pend", &pst3, g_prof_rp_pend);
+            /* %llu, never %zu: msvcrt printf (Windows) lacks %zu and
+             * newer MinGW-w64 toolchains reject it under -Werror */
             fprintf(stderr, "[prof] loop iters=%llu poll0=%llu "
-                    "n=%zu pfn=%zu dirty=%d\n",
+                    "n=%llu pfn=%llu dirty=%d\n",
                     (unsigned long long)g_prof_rp_iters,
                     (unsigned long long)g_prof_rp_poll0,
-                    n, pf_n, (int)pf_dirty);
+                    (unsigned long long)n, (unsigned long long)pf_n,
+                    (int)pf_dirty);
         }
 
         for (size_t i = 0; i < n; i++) {
